+++ Portable/RCS.PortableShop.ServiceClients.Products/Wrappers/ServiceOptions.cs
@@ -1,0 +1,7 @@
+﻿namespace RCS.PortableShop.ServiceClients.Products.Wrappers
+{
+    public class ServiceOptions
+    {
+        public string RemoteAddress { get; set; }
+    }
+}